% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{random_invertable_matrix}
\alias{random_invertable_matrix}
\title{Generate a random invertible matrix of integers}
\usage{
random_invertable_matrix(n, seed = NA)
}
\arguments{
\item{n}{Number of rows and columns in the square matrix.}

\item{seed}{A character string from which a seed for the random number generator will be generated.
You can use your user id or email address to get a reproducible matrix.}
}
\description{
This function generates random square integer matrices.
It tests to be sure the random matrix is invertible (does not have a determinant of zero);
if the determinant is zero, it keeps trying until it produces a solvable matrix.
}
\examples{
random_invertable_matrix(3, "bob")
}
\keyword{invertible}
\keyword{matrix}
\keyword{random}

