% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{rotation_matrix}
\alias{rotation_matrix}
\title{Generate a rotation matrix to transform a 2-column input by rotating it through a given angle.}
\usage{
rotation_matrix(theta)
}
\arguments{
\item{angle}{in radians}
}
\description{
Generate a rotation matrix to transform a 2-column input by rotating it through a given angle.
}

