\docType{package}
\name{tyr}
\alias{tyr}
\alias{tyr-package}
\title{Utility Functions for Interactive R Sessions}
\description{
Miscellaneous Utility Functions for Interactive R Sessions
}
\details{
\tabular{ll}{ Package: \tab tyr\cr Type: \tab Package\cr
Version: \tab 1.0\cr Date: \tab 2014-02-21\cr License: \tab
GPL (>= 3)\cr LazyLoad: \tab yes\cr }

Convenience functions for common operations in interactive
R sessions
}
\author{
Tyler Smith \email{tyler@plantarum.ca}
}
\keyword{correlation}
\keyword{package}

